Change component dependencies and messaging dispatch structure
Initial understanding of connections between MQTT, MessageService and API was flawed.
Message Service should in fact not be tied in any tight way to MQTT, other than if messages
are sent to/from MQTT, then the message dispatch functions should take that into account, e.g.
by picking out the MQTT manager instance as an injected key assoc-ed to the message it should
do something with.

Now, the set up has changed to the following:
- MessageService depends on none
- MQTT depends on MessageService, and will pass itself along with message to MessageService
- API depends on both MessageService and MQTT, although it might only need to depend on MQTT, since
it already has the MessageService dependency.

The Message Service is now set up with handlers spread out under the 'platform.message.dispatch.*' namespace.
The reason is to keep it modular and tidy. There is also a message dispatch interface, which is basically the
multimethod definition all other message dispatch functionality should add to.