Add a room to the database
Conceptually a room is mean to be a place where a user can invite
another user to play with them in the room. There are a lot of ideas
on how the concept can be used, but the current implementation is bare
bones.

The tables are set up for future expansion.