Add FX table to SQL database
This table is heavily inspired by the visual user interface, and how
musicians think about this (based on very old software, which in turn
is based on very old physical hardware and their layouts). You have a
number of presets that you can choose in typical music software, that
hold the values for your application that you change depending on
which instrument you are playing.

In this jargon used by musicians, a slot represents as a physical slot
where you store your settings (but only for the specific task at
hand). As such you have a LIMITED amount of slots at hand, they are
not named, but rather numbered and have a physical hierarchy (top to
bottom, or left to right) in the interface. This is represented in the
table with the slot column (a simple integer).

We also need to save the currently used FX settings, which do not
reflect any preset. Here we simply reuse the table and make the number
0 mean the current FX settings (which always are updated to reflect
the latest setting).

Each preset is at its core a key/value combo, where the key controls
which FX is changed, and the value controls how much the effect is
affecting the real world.

The values in the presets themselves are stored in two tables: fx_key and
fx_value. This could have been represented as a JSONB column in
fx_preset. I opted out of this because my experience with using JSONB
as storage is that you lose the schema, and it's now up to the
application to enforce it. This has always led to tears. The option
was to split this up in the database, and have certain things enforced
via the db schema. The main table used to enforce this is fx_key,
which has to hold the effects supported.

The way it's been setup will also allow for database enforced
differations between which FX are allowed for different types of
Teleporters. This is not needed right now, but given physical hardware
it's a distinct possibility, and so it's a small cost to setup in this stage.