Add authn/authz to routes
The REST APIs under /api now needs authentication. No authz has yet
been added.

Added redis for session handling, and updated the server to initialize
with a redis store, a authn/z backend and the postgres database.

In addition, changed the api endpoints for auth to use the postgres
database, instead of the in-memory db for the jam handling.