Add profile API
Added profile model, api and HTTP routes.

Manual testing for the base64 image has been done, and it seems to
work. REST API endpoints were added for getting a profile, saving a
profile and getting the pronouns that the solution supports.

Currently there is no backoffice, and as such the pronouns will have
to be manually added. -1 is the default pronoun (which is a blank text).