Add room session history
Who played with whom, when and in which room has been added. There are
a lot of potential edge cases in an implementation such as this, so I
have chosen to minimize the surface area by being as dumb as
possible. It should provide some understandable defaults that users
can work with.