Remodel RoomDB to handle MQTT and Jam in addition
RoomDB now becomes the central point for handling all logic with
regards to the Room Sessions.

The API is split up in two parts: One REST API, and one MQTT API. The
MQTT API has not yet been hooked up with the frontend, and as such
might see further changes.

Jam is prepared for, but not tested with live devices.